Working condition until mega million
