+++ app/src/main/java/com/example/lotterypickerkotlin/ui/main/TopTabSectionsAdapter.kt
@@ -8,6 +8,10 @@ import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.lotterypickerkotlin.R
+import com.example.lotterypickerkotlin.fragments.Cash5Fragment
+import com.example.lotterypickerkotlin.fragments.Pick3Fragment
+import com.example.lotterypickerkotlin.fragments.Pick4Fragment
+import java.lang.reflect.Array.newInstance
 
 private val TAB_TITLES = arrayOf(
     R.string.tab_text_1,
@@ -24,7 +28,7 @@ class TopTabSectionsAdapter(
     fm: FragmentManager,
     lifecycle: Lifecycle
 ) : FragmentStateAdapter(fm, lifecycle) {
-
+    val testList = mutableListOf<Fragment>()
     val index = Utils.index
 
     fun getPageTitle(position: Int): CharSequence {
@@ -36,6 +40,11 @@ class TopTabSectionsAdapter(
     }
 
     override fun createFragment(position: Int): Fragment {
-        return PlaceholderFragment.newInstance(index)
+      return when(position) {
+          0 -> Pick3Fragment()
+          1 -> Pick4Fragment()
+          2 -> Cash5Fragment()
+          else -> Pick3Fragment()
+      }
     }
 }
